add dropDown & rename icon components
