+++ src/assets/icons/check-email-icon.tsx
