+++ src/assets/icons/clean-input.tsx
