+++ src/assets/icons/edit-text.tsx
