+++ src/assets/icons/log-out.tsx
