+++ src/assets/icons/watch-pass-crossed-out.tsx
