+++ src/components/auth/personal-info/personal-info.tsx
@@ -4,7 +4,7 @@ import { clsx } from 'clsx'
 
 import s from './personal-info.module.scss'
 
-import { EdittextIcon } from '@/assets/icons/editText.tsx'
+import { EdittextIcon } from '@/assets/icons/edit-text'
 import { ProfileDisplayMode } from '@/components/auth/personal-info/display-mode'
 import { ProfileEditMode } from '@/components/auth/personal-info/edit-mode'
 import { Avatar, Button, Card, Typography } from '@/components/ui'
